Remove profile saved alert; move LanguageOptionViewModel
Removed ProfileSaved event and alert from settings. Moved LanguageOptionViewModel to its own file and namespace. Updated export_pdf_en.png. Minor formatting changes.
